+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v1.2](https://github.com/sigfox-tech-radio/sigfox-ep-rf-api-semtech-sx126x/releases/tag/v1.2) - 22 Mar 2024
+
+### Fixed
+
+* SX1262 **hp_max** configuration issue
+* BIDIRECTIONAL flag compilation issue
+
 ## [v1.1](https://github.com/sigfox-tech-radio/sigfox-ep-rf-api-semtech-sx126x/releases/tag/v1.1) - 15 Dec 2023
 
 ### General